Convert `checksum_verification` test to spec.
